Refactor Order and all subclasses to support save/load
The new pattern is to never directly call the constructor for Orders and
Order subclasses. Instead call a `OrderClass.create_order_class` factory
method which creates the object for you and initializes it. This allows
the constructor to not have to touch Entities or other objects that
might not be loaded when this Order object is loaded.

That's important because we load the Orders serially and need to have a
partially initialized object so we can later reference it when finishing
setting up Orders.

We'll use the same pattern for Effects and SensorImages. All of these
things are loaded in-line with Entitites and they all reference each
other so might run into the same problems. This follows the same pattern
of Entities never requiring another Entity in the contructor and instead
relying on post-object-creation setup to handle any inter-Entity
references.